+++ backend/handlers/markets/getmarkets.go
@@ -1,5 +1,5 @@
 package marketshandlers
 
 type PublicMarket struct {
-	MarketID uint `json:"marketId"`
+	MarketID int64 `json:"marketId"`
 }
